worked on test data

+++ recsys/models.py
@@ -51,4 +51,4 @@ class CosineSimilarity(models.Model):
     created_time = models.DateTimeField('created_time', blank=True, null=True)
     source_id = models.CharField(max_length=100, blank=True, null=True) #characters id of post
     target_id = models.CharField(max_length=100, blank=True, null=True)
-    similarity = models.DecimalField(max_digits=5, decimal_places=5, blank=True, null=True)
+    similarity = models.FloatField(blank=True, null=True)
